Refactor data (#103)
