+++ fast_llm/data/dataset/gpt/fim/config.py
@@ -1,0 +1,58 @@
+from fast_llm.config import Config, Field, FieldHint, check_field, config_class, skip_valid_if_none
+from fast_llm.utils import Assert
+
+FIM_PREFIX = "<fim_prefix>"
+FIM_MIDDLE = "<fim_middle>"
+FIM_PAD = "<fim_pad>"
+FIM_SUFFIX = "<fim_suffix>"
+
+
+@config_class()
+class FimConfig(Config):
+    """
+    Configuration for FIM.
+    """
+
+    rate: float = Field(
+        default=0.0,
+        desc="FIM rate for each sample.",
+        hint=FieldHint.core,
+        valid=check_field(Assert.in_range_incl, 0, 1),
+    )
+    max_middle_len: int | None = Field(
+        default=None,
+        desc="Maximum length of the middle segment in FIM.",
+        hint=FieldHint.feature,
+        valid=skip_valid_if_none(check_field(Assert.gt, 0)),
+    )
+    split_sample: str | None = Field(
+        default=None,
+        desc="Split samples on this token and permute each fragment separately.",
+        hint=FieldHint.feature,
+    )
+    fragment_rate: float = Field(
+        default=0.0,
+        desc="FIM rate for each fragment when using fim_split_sample.",
+        hint=FieldHint.feature,
+        valid=check_field(Assert.in_range_incl, 0, 1),
+    )
+    ignore_prefix: str | None = Field(
+        default=None,
+        desc="Do not apply FIM to fragments that start with this prefix.",
+        hint=FieldHint.feature,
+    )
+    spm_rate: float = Field(
+        default=0.5,
+        desc="TODO.",
+        hint=FieldHint.feature,
+        valid=check_field(Assert.in_range_incl, 0, 1),
+    )
+    truncate_or_pad: bool = Field(
+        default=False,
+        desc="TODO.",
+        hint=FieldHint.feature,
+    )
+
+    def _validate(self):
+        super()._validate()
+        Assert.in_range_incl(self.rate, 0, 1)
