+++ fast_llm/data/dataset/gpt/memmap.py
@@ -4,7 +4,7 @@ import typing
 
 import numpy as np
 
-from fast_llm.data.gpt.dataset import GPTIndexedDataset
+from fast_llm.data.dataset.gpt.abstract import GPTIndexedDataset
 from fast_llm.data.preparator.gpt_memmap.config import MEMMAP_DTYPES, MEMMAP_DTYPES_INV, MEMMAP_INDEX_HEADER
 from fast_llm.engine.config_utils.data_type import DataType
 from fast_llm.utils import Assert, div
