+++ fast_llm/data/dataset/gpt/slice.py
@@ -1,5 +1,5 @@
-from fast_llm.data.config import SamplableSplitDataset
-from fast_llm.data.gpt.dataset import GPTIndexedDataset
+from fast_llm.data.dataset.abstract import SamplableSplitDataset
+from fast_llm.data.dataset.gpt.abstract import GPTIndexedDataset
 from fast_llm.engine.distributed.config import PhaseType
 from fast_llm.utils import Assert, normalize_probabilities, padded_cumsum
